Add a bit more heuristic verification before blindy installing to a directory that might be wrong.
- This code currently lacks the context of "Am I installing a Packages.config package?"
- A project might have a packages directory, but not used for the package being linked.
- Now, require that the package exists in that directory before executing an install to it.

+++ packages/lib/src/installer.ts
@@ -61,7 +61,10 @@ export function installPackages({
 			}
 		});
 
-		if (fs.existsSync(localPackagesDirectory)) {
+		if (
+			fs.existsSync(localPackagesDirectory) &&
+			fs.readdirSync(localPackagesDirectory).some(x => x.startsWith(name))
+		) {
 			const installLocalCommand = `install ${name} -Version ${version} -Source ${directory} -OutputDirectory ${localPackagesDirectory}`;
 			log(`Exec: nuget ${installLocalCommand}`);
 			execNuget(installLocalCommand, {
